8706 move patches to proper paths / update comments
