+++ lib/pg_saurus/connection_adapters.rb
@@ -4,7 +4,6 @@ module PgSaurus::ConnectionAdapters # :nodoc:
   autoload :AbstractAdapter
   autoload :PostgreSQLAdapter,  'pg_saurus/connection_adapters/postgresql_adapter'
   autoload :Table
-  autoload :IndexDefinition,    'pg_saurus/connection_adapters/index_definition'
   autoload :FunctionDefinition, 'pg_saurus/connection_adapters/function_definition'
   autoload :TriggerDefinition,  'pg_saurus/connection_adapters/trigger_definition'
 end
